DO NOT MERGE. Fix crash in STK app when launching browser
The problem was that StkAppService.launchBrowser() was calling

        intent.setClassName("com.android.browser",
                "com.android.browser.BrowserActivity");

when launching the browser.  But in ICS, the browser that's actually
installed is from the package com.google.android.browser, not
com.android.browser.  So right now, the STK crashes if it ever tries to
view a web page.

The fix is for the STK app to just not hardcode the browser package/class
name at all, since if you fire off a VIEW intent with an http: URI in the
data, you'll get whatever the default browser is.  Also, if the STK gets a
request with no URI at all, we bring up http://google.com/ as a default
"home page".

Bug: 5489975
Change-Id: If3f9d1468562a5e7f948156aa8395525665bc9ff