StkAppSerice: Stop the service when CatService Instance is null
ANR is seen in StkAppService because of ServiceHandler is not
created due to CatService is not running.

Stop Stkapp Service if Catservice is not running to avoid ANR.
Whenever service is started using StartService first onCreate
followed by onStart will be called. Next time onwards only
Onstart will be called due to this ServiceHandler will not intialize.
Hence moving corresponding code to onStart from onCreate.

Change-Id: I5a0a78d2012e3bf5aff2b5512dd21aa7bb96ec4a