stk: Add support of Help information
According to the 3GPP TS 31.124, Help
information is an optional feature. If help
information is set as supported, then the
proactive command should support Help menu.
If Help menu is supported but not available
then conformance tests cases are considered
as failed(eg: 27.22.4.8.2 sequence 2.1).

This patch adds support for the optional
feature - Help Info.

Change-Id: I27412558853ff2f23b83a924e14efdf2de8d285f
Author: Guillaume Lucas <guillaumex.lucas@intel.com>
Signed-off-by: Guillaume Lucas <guillaumex.lucas@intel.com>
Signed-off-by: Nizar Haouati <nizar.haouati@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 2883