Changed timeouts for Stk
The SIM Toolkit interface in Android uses one common
timeout for UI commands. This is not good practice,
as there is a very clear difference between use cases.

A DISPLAY TEXT command with the option “clear after delay”
and with option “Wait for user to clear” should not have
the same timeout, which is the case without a patch. The
timeouts have in the patch been set to 15 s for
“clear after delay” (i.e. automatically removed), 60 seconds
for “Wait for user to clear” and 30 s for other UI timeouts.

The times are both in accordance with ETSI 102.223 and
operator requirements.

Change-Id: I61262bf36a84f071ec4f223eb187f92e2026b68b