Stk: Fix issue in DISPLAY TEXT-Sustained text
3GPP TS 31.124 test case 27.22.4.1.4/1 Display Text -
sustained text is failing.

As per the CAT specification, text shall
be sustained until a subsequent proactive command
is received containing display data. Existing
dialog showing the text is not replaced with
the newly received text message resulting in
conformance case failing.

waitForUserResponse not set correctly and also
Intent.FLAG_ACTIVITY_MULTIPLE_TASK is used
resulting in a starting of a new task even if there
is an existing task.

Fix the issue by setting the waitForUserResponse
to false on response not needed and on screen
not available. Also, Intent.FLAG_ACTIVITY_CLEAR_TASK
is used as one of the flags in creating the intent
for launching the dialog.

Change-Id: Ic614f96883aaa92696982ca41d85a02341bcb132
Author: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 61998