Stk notifications has purple background
Code was introduced in frameworks/base
96fd7c1c1acc03b40b1813ef65793560c175ef80 that makes notifications
refering to status_bar_latest_event_content.xml directly show up
in purple.

Using internal resource directly is discouraged, or at least the author
bids the user of this directly to have fun with the purple background.
The recommended way is to use the class Notification.Builder (Api 11)
to build your notifications. This commit uses Notification.Builder
class which setup the correct notification layout and background.

Change-Id: Idd850a16b7b68a40726a21947adc58cdcdfa1183