Improve STK idle mode text handling
The STK idle mode text was not shown as a ticker text in the idle
mode screen, but only when you dragged down the notification bar
to read the notification. Moreover, the text was shown without any
title which made the context of this notification unclear.

Now it shows the STK title text as specified by the SIM card.

Change-Id: I0388bebad5d9da26e934a5552aa19d876cc5405c