STK: prevent to appear a home screen when DISPLAY_TEXT
This patch changes the acvity flag as FLAG_ACTIVITY_CLEAR_TOP from
FLAG_ACTIVITY_CLEAR_TASK. FLAG_ACTIVITY_CLEAR_TASK destroys all
activities and creates a new activity. The home screen is displayed
when DISPLAY_TEXT is occured and another STK activity is already performing.
If FLAG_ACTIVITY_CLEAR_TOP is used, another STK activity is brought to
foreground and then it is cleared. So, the home screen is not appeared.

Bug: 10316520
Change-Id: I19a9fb37e091c30e0094b7734fbb5e406aae47f2