Add support for idle screen notification
When the device is not in an interactive state or
when the screen off intent is received from the framework,
we can assume that the screen is idle.

Add support for idle screen available set up event list
and event download.

Handle normal priority and high priority display text
correctly.

Launch the Idle Mode Text when the idle screen notification
is received.

Bug: 18274506
Change-Id: Iff69fd8f998746a0fe37577b2bb63086478ef448