The dialog shall be displayed even if the command comes while sleeping
This change fixes the issue that StkDialogActivity is unexpectedly
finished before user sees it. The issue can happen if the terminal
receives DISPLAY TEXT command with the immediate response option while
the terminal is sleeping.

StkAppService has a mechanism to finish the pending dialog when the
current session ends. The state of the activity immediately goes to the
stop state in this use-case and the session also ends right away because
of the immediate response option. The dialog shall not be treated as
pending dialog if the immediate response option is specified.

Change-Id: I7a6e87e24d91bb91e1708bc28e5aab69f1cb506c