Specify the task affinity for ToneDialog
As there is no task affinity specified for ToneDialog, only ToneDialog
can be separated from the other activities in the STK application. For
example, we can see StkMenuActivity and ToneDialog as separated tasks in
the split-screen mode of the multi-window feature. That should not be a
designed behavior. It must be better to apply the same task affinity as
the other ones.

Bug: 32652551
Test: manual - confirmed that ToneDialog is not separated from others.

Change-Id: I0f5cc21807a61de320e2b15f3f6a79c4f3f651df