Wake up from the standby mode and display message on the keyguard screen
If the device is in the standby mode when receiving DISPLAY TEXT command
or GET INPUT/INKEY command from the SIM card, it must get out of the
standby mode, display a message on the keyguard screen. Once the device
is unlocked, it must display the screen or the dialog associated with
the command. Refer to the descriptions in GSMA PDATA.12.

Bug: 70004658
Test: Performed manual test cases for the affected scenarios.

Change-Id: Ieeafd446e89882b2af163677c67db5f734d1cd9d