Pending secondary menu shall remain in the background
Secondary menu shall remain in the background while SIM Toolkit is
displaying a dialog on it, but it disappears if user goes to the other
application and comes back to SIM Toolkit. To fix that issue,
getAppTasks() shall be used instead of getRunningTasks() which can
return some other tasks not associated with the calling application.

Bug: 112742596
Test: Confirmed that the pending activity can remain in the use-case.

Change-Id: Id268a0b68a2fb6ba226d87cd054cdf5dbe3e8ea4