Fix terminal response not sent during sim hotswap.
In a case, where single sim is removed and reinserted
terminal response for the subsequent display text/select
item proactive command is not sent resulting in GCF test
case failure.

Fix: When sim card is absent initialize the CatService instance
to null., So that StkAppService always use the latest CatService
instance when sim card becomes present.

Bug: 74049285
Change-Id: I48e403354a04e5d2757b47202706ad461ad242aa