Make it configurable to display the allowed characters in the digit mode
According to ETSI TS 102 223, the digit only mode of GET INKEY/INPUT
commands allow user to enter not only the numeric characters 0 to 9 but
also some symbol characters, i.e. *, # and +. Such allowed characters
are displayed on the input screen in the current implementation, but it
is quite confusing when only the numeric characters are used for the
actual services in the market. It must be left to service provider
whether or not and how to explain the allowed characters to user.

Bug: 128890104
Test: Confirmed the expected behavior in manual test cases.

Change-Id: Id254f5d1d816813b84235db47fbb37c98cb69a32