Treat recentapps as home press for STK dialogs
For interactive STK dialogs which listen for user presses of the HOME
key, we will treat RECENTAPPS presses as equivalent. The reason for this
is because with new gesture-based navigation it's hard for the user to
trigger a HOME press, and the same user intention often triggers a
RECENTAPPS press.

Bug: 161207764
Test: manual
Change-Id: Ib64068f191e62b7103472d39525305be3dab59c8
Merged-In: Ib64068f191e62b7103472d39525305be3dab59c8